requirements.txt
Flask==1.0.2
boto3==1.7.65
Flask-SQLAlchemy==2.3.2
